User Testing Complete
Tested about page on different users, feed back was negative about the height changes between tabs and the jerky-ness this caused in the browser. The solution was to fix the height of all tabs.
HTML structure adjusted to allow animation to appear to be coming from the center of the box